requirements.txt
PyDispatcher==2.0.5
scrapy==2.6.1
